+++ dash/constants.py
@@ -1,9 +1,14 @@
+# COLUMN TYPES
+NUMERICAL_COLUMN = "numerical"
+CATEGORICAL_COLUMN = "categorical"
+
 # ADESIT COLUMNS IN DATASET
 ADESIT_INDEX = "id"
 G12_COLUMN_NAME = "_violating_tuple"
 G3_COLUMN_NAME = "_g3_to_remove"
 SELECTION_COLUMN_NAME = "_selected"
 PROJ_AXES = ['__PROJ1', '__PROJ2']
+SUFFIX_OF_ENCODED_COLS = "__LEADESIT"
 
 # MAX NUMBER OF ROWS IN DATATABLES
 TABLE_MAX_ROWS = 16
