+++ dash/utils/heatmap_utils.py
@@ -1,14 +1,20 @@
 import numpy as np
 
 import plotly.graph_objects as go
-from utils.figure_utils import gen_subplot_fig, adjust_layout, add_basic_histograms, add_advanced_histograms, convert_from_numpy_edges
+from utils.figure_utils import gen_subplot_fig, adjust_layout, convert_from_numpy_edges
+import utils.histogram_utils as hist_gen
 
-def heatmap_basic_bloc(fig, df, xaxis_column_name, yaxis_column_name, resolution):
+from constants import *
+
+def heatmap_basic_bloc(fig, df, xaxis_column_name, yaxis_column_name, resolution, session_infos):
     colorscale=[
         [0.0, 'rgba(0,0,0,0)'],
         [1.0, 'rgba(0,0,0,1)'],
     ]
-
+    
+    # We need to get the label encoded columns if the axis is categorical
+    if session_infos["user_columns_type"][xaxis_column_name]=="categorical": xaxis_column_name += SUFFIX_OF_ENCODED_COLS
+    if session_infos["user_columns_type"][yaxis_column_name]=="categorical": yaxis_column_name += SUFFIX_OF_ENCODED_COLS
     H, xedges, yedges = np.histogram2d(df[xaxis_column_name], df[yaxis_column_name], bins=(resolution, resolution))
     fig.add_trace(
         go.Heatmap(
@@ -44,19 +50,19 @@ def heatmap_basic_bloc(fig, df, xaxis_column_name, yaxis_column_name, resolution
 
 def basic_heatmap(df, xaxis_column_name, yaxis_column_name, resolution, session_infos):
     fig = gen_subplot_fig(xaxis_column_name, yaxis_column_name)
-    fig = heatmap_basic_bloc(fig, df, xaxis_column_name, yaxis_column_name, resolution)
-    fig = add_basic_histograms(fig, df, xaxis_column_name, yaxis_column_name, resolution, session_infos)
+    fig = heatmap_basic_bloc(fig, df, xaxis_column_name, yaxis_column_name, resolution, session_infos)
+    fig = hist_gen.add_basic_histograms(fig, df, xaxis_column_name, yaxis_column_name, resolution, session_infos)
     fig = adjust_layout(fig, df, xaxis_column_name, yaxis_column_name, session_infos)
     return fig
 
 def advanced_heatmap(df, label_column, xaxis_column_name, yaxis_column_name, resolution, session_infos):
     fig = gen_subplot_fig(xaxis_column_name, yaxis_column_name)
-    fig = heatmap_basic_bloc(fig, df, xaxis_column_name, yaxis_column_name, resolution)
+    fig = heatmap_basic_bloc(fig, df, xaxis_column_name, yaxis_column_name, resolution, session_infos)
     
     non_problematics_df = df.loc[df[label_column] == 0]
     problematics_df = df.loc[df[label_column] > 0]
 
-    fig = add_advanced_histograms(fig, non_problematics_df, problematics_df, xaxis_column_name, yaxis_column_name, resolution, session_infos)
+    fig = hist_gen.add_advanced_histograms(fig, non_problematics_df, problematics_df, xaxis_column_name, yaxis_column_name, resolution, session_infos)
     fig = adjust_layout(fig, df, xaxis_column_name, yaxis_column_name, session_infos)
     fig.update_layout(barmode='group')
     fig.update_layout(barmode='group')
