+++ deploy/requirements.txt
@@ -5,14 +5,9 @@ dash-tabulator==0.4.2
 flask_caching==1.10.1
 pandas==1.3.5
 pydataset==0.2.0
-scikit-learn==0.24.2
+scikit-learn==1.0.2
 dash-cytoscape==0.3.0
 prince==0.7.1
 
 numpy==1.22.1
 cython==0.29.23
-
-# sortedcontainers==2.4.0
-# networkx==2.5
-# recordlinkage==0.14
-# scikit-learn==0.24.1
